feat: finalize cropping logic
